backend/requirements.txt
asyncpraw
python-dotenv
sqlalchemy
asyncpg
google-genai
schedule